backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
twilio==8.12.0
openai==1.30.0
anthropic==0.28.0
google-cloud-texttospeech==2.16.3
SQLAlchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0
pyngrok==7.1.6
textblob==0.18.0
python-multipart==0.0.9
